+++ main/src/main/java/net/oneandone/stool/configuration/Autoconf.java
@@ -75,7 +75,7 @@ public class Autoconf {
             dflt = dest.defaults.get("");
             dflt.put("maven.opts", "-Xmx1024m -Dmaven.repo.local=@localRepository@ @trustStore@");
             dflt.put("template", "ciso-tomcat");
-            dflt.put("container.env", "mode:test");
+            dflt.put("template.env", "mode:test,fault:false");
             dest.defaults.put("svn:https://svn.1and1.org/svn/controlpanel_app/controlpanel/", cp());
             dest.defaults.put("svn:https://svn.1and1.org/svn/sales/workspaces/", workspace());
         }